+++ failover-core/src/main/java/com/societegenerale/failover/core/observable/DefaultFailoverObserver.java
@@ -78,7 +78,7 @@ public class DefaultFailoverObserver implements FailoverObserver {
         failoverScanner.findAllFailover().forEach(failover->  {
             Metrics metrics = Metrics.of("failover-report-%s".formatted(failover.name()))
                     .collect("name", failover.name())
-                    .collect("expiry-duration", Long.toString(failoverExpiryExtractor.expiryDuration(failover)))
+                    .collect("expiry-duration", failoverExpiryExtractor.expiryDuration(failover))
                     .collect("expiry-unit", failoverExpiryExtractor.expiryUnit(failover).name());
             genericInfo.forEach(metrics::collect);
             observablePublisher.publish(metrics);